优化一些代码

+++ hv/utils/memory_utils.h
@@ -18,7 +18,7 @@ namespace utils
 
 		NTSTATUS free_user_memory(_In_ HANDLE process_id, _In_  PVOID base_address, _In_ SIZE_T size,  bool hide = true);
 
-		NTSTATUS lock_memory(unsigned long long  address, size_t size, OUT PMDL* out_mdl);
+		NTSTATUS lock_memory(unsigned long long  address, ULONG size, OUT PMDL* out_mdl);
